refactor(skills): extract source test modules
- Move skill unit tests into sibling *_test.rs files
- Normalize existing skills test module filenames
