+++ crates/aion-skills/src/frontmatter_supplemental_test.rs
@@ -1,5 +1,5 @@
 // Supplemental tests for frontmatter.rs — covers test-plan.md cases not in impl tests.
-// Included from frontmatter.rs as: #[cfg(test)] mod supplemental_tests;
+// Included from frontmatter_test.rs as frontmatter_supplemental_test.
 // `use super::*` gives access to private functions in frontmatter.rs.
 
 use super::*;
