+++ crates/aion-skills/src/loader_test.rs
