+++ crates/aion-skills/src/shell_supplemental_test.rs
